+++ ci/test/06_script_b.sh
@@ -69,6 +69,7 @@ if [ "${RUN_TIDY}" = "true" ]; then
           " src/util/bytevectorhash.cpp"\
           " src/util/check.cpp"\
           " src/util/error.cpp"\
+          " src/util/exception.cpp"\
           " src/util/getuniquepath.cpp"\
           " src/util/hasher.cpp"\
           " src/util/message.cpp"\
