+++ src/qt/guiutil.cpp
@@ -20,6 +20,7 @@
 #include <protocol.h>
 #include <script/script.h>
 #include <script/standard.h>
+#include <util/exception.h>
 #include <util/system.h>
 #include <util/time.h>
