+++ src/util/system.h
@@ -19,12 +19,10 @@
 #include <fs.h>
 #include <logging.h>
 #include <sync.h>
-#include <tinyformat.h>
 #include <util/settings.h>
 #include <util/time.h>
 
 #include <any>
-#include <exception>
 #include <map>
 #include <optional>
 #include <set>
@@ -51,8 +49,6 @@ bool error(const char* fmt, const Args&... args)
     LogPrintf("ERROR: %s\n", tfm::format(fmt, args...));
     return false;
 }
-
-void PrintExceptionContinue(const std::exception* pex, std::string_view thread_name);
 
 /**
  * Ensure file contents are fully committed to disk, using a platform-specific
