+++ src/util/thread.cpp
@@ -5,7 +5,7 @@
 #include <util/thread.h>
 
 #include <logging.h>
-#include <util/system.h>
+#include <util/exception.h>
 #include <util/threadnames.h>
 
 #include <exception>
